feat(FrontEnd): Boats list view done

+++ Backend/FishingBooker/FishingBooker/src/main/java/com/isa/FishingBooker/service/interfaces/RevisionService.java
@@ -8,7 +8,7 @@ import com.isa.FishingBooker.model.revision.TutorServiceAppointmentRevision;
 
 public interface RevisionService extends Service<Revision> {
 	public void updateRevisionStatus(int revisionId, Status revisionStatus, int adminId);
-	public int getBoatAppointmentRevisionsRate(Integer id);
+	public double getBoatAppointmentRevisionsRate(Integer id);
 	public int getResortAppointmentRevisionsRate(Integer id);
 	public double getTutorServiceAppointmentRevisionsRate(Integer id);
 	public String makeTutorServiceRevision(TutorServiceAppointmentRevision revision);
